requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.0
billiard==3.6.4.0
black==21.6b0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.7
Django==3.2.4
django-allauth==0.44.0
django-celery-beat==2.2.1
django-decouple==2.1
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.1
django-filter==2.4.0
django-nine==0.2.4
django-otp==1.0.6
django-phonenumber-field==5.2.0
django-redis==5.0.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.7.2
elasticsearch==7.13.3
elasticsearch-dsl==7.3.0
idna==2.10
kombu==5.1.0
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.8.1
phonenumbers==8.12.26
prompt-toolkit==3.0.19
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyotp==2.6.0
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
